requirements.txt
mysql_connector_repackaged==0.3.1
PyMySQL==1.1.1
PyQt5==5.15.11
PyQt5_sip==12.17.0
PyQt_Fluent_Widgets==1.7.6
tabulate==0.9.0
